+++ src/solvers/mg/MGScatterFissionOperator.cc
@@ -41,7 +41,7 @@ MGScatterFissionOperator(SP_input         input,
   // Number of active groups
   int upper = d_number_groups;
   if (d_adjoint) upper = -1;
-  d_number_active_groups = std::abs(upper - d_group_cutoff);
+  d_number_active_groups = std::abs(upper - (int)d_group_cutoff);
 
   d_moments_size = d_mesh->number_cells();
